f1

+++ baseproject/cli.py
@@ -2,13 +2,23 @@ from importlib.resources import files
 from pathlib import Path
 import shutil
 from django.core.management.utils import get_random_secret_key
+import argparse
+import sys
 
 def create_project(name):
+    parser = argparse.ArgumentParser(description="Create Django base project")
+    parser.add_argument("name", help="Project name")
+
+    args = parser.parse_args()
+    name = args.name
+
     template_dir = files("baseproject").joinpath("template")
 
-    shutil.copytree(template_dir, name)
 
-    # Create .env
+    if not Path(template_dir).exists():
+        raise FileNotFoundError(f"Template folder not found at {template_dir}")
+
+     # Create .env
     env_path = files(name, ".env")
     with open(env_path, "w") as f:
         f.write(f"DJANGO_SECRET_KEY={get_random_secret_key()}\n")
@@ -19,8 +29,10 @@ def create_project(name):
         f.write(f"EMAIL_HOST_USER=example@gmail.com\n")
         f.write(f"EMAIL_HOST_PASSWORD=**** **** **** ***\n")
 
+    if Path(name).exists():
+        print(f"Error: Folder '{name}' already exists.")
+        sys.exit(1)
+
     shutil.copytree(str(template_dir), name)
 
     print(f"Project '{name}' created successfully.")
-
-
